Update Classes depending on NEAT
